requirement.txt
Django==2.0.1
django-extensions==1.9.9
mysqlclient==1.3.12
pytz==2017.3
requests==2.18.4
djangorestframework==3.8.2